7.pizza_redis_pipeline/sql-mondo-redis-kafka/redis_start/requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
confluent-kafka==2.13.0
dnspython==2.8.0
idna==3.11
mysql-connector-python==9.6.0
psutil==7.2.2
pymongo==4.16.0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.1
requests==2.32.5
urllib3==2.6.3
